rename ./server/ to ./pds/
